Assignment No 11

+++ Assignment No. 10/05 Write a program which accept area in square feet and convert it into square meter.c
@@ -14,7 +14,7 @@ double SquareMeter(int iValue)
     }
     else
     {
-        return (double)( iValue) * 0.0929;
+        return iValue * 0.0929;
     }
 }
 int main()
